+++ lib/cldr/calendar/backend/week_compiler.ex
@@ -282,8 +282,18 @@ defmodule Cldr.Calendar.Compiler.Week do
         {year, month, day, hour, minute, second, microsecond}
       end
 
+      def add(year, month, day, hour, minute, second, microsecond, :quarter, step) do
+        {year, month, day} = plus(year, month, day, :quarters, step)
+        {year, month, day, hour, minute, second, microsecond}
+      end
+
       def add(year, month, day, hour, minute, second, microsecond, :month, step) do
         {year, month, day} = plus(year, month, day, :months, step)
+        {year, month, day, hour, minute, second, microsecond}
+      end
+
+      def add(year, month, day, hour, minute, second, microsecond, :week, step) do
+        {year, month, day} = plus(year, month, day, :weeks, step)
         {year, month, day, hour, minute, second, microsecond}
       end
 
@@ -385,6 +395,7 @@ defmodule Cldr.Calendar.Compiler.Week do
       Implements the `Inspect` protocol for `Date` in this calendar
       """
       calendar_impl()
+
       @spec inspect_date(Calendar.year(), Calendar.month(), Calendar.day(), Inspect.Opts.t()) ::
               Inspect.Algebra.t()
       def inspect_date(year, month, day, _) do
@@ -395,6 +406,7 @@ defmodule Cldr.Calendar.Compiler.Week do
       Implements the `Inspect` protocol for `DateTime` in this calendar
       """
       calendar_impl()
+
       @spec inspect_datetime(
               Calendar.year(),
               Calendar.month(),
@@ -446,6 +458,7 @@ defmodule Cldr.Calendar.Compiler.Week do
       Implements the `Inspect` protocol for `NaiveDateTime` in this calendar
       """
       calendar_impl()
+
       @spec inspect_naive_datetime(
               Calendar.year(),
               Calendar.month(),
@@ -466,6 +479,7 @@ defmodule Cldr.Calendar.Compiler.Week do
       Implements the `Inspect` protocol for `Time` in this calendar
       """
       calendar_impl()
+
       @spec inspect_time(
               Calendar.hour(),
               Calendar.minute(),