+++ lib/cldr/calendar/base/week.ex
@@ -214,6 +214,11 @@ defmodule Cldr.Calendar.Base.Week do
     date_from_iso_days(iso_days, config)
   end
 
+  def add(naive_datetime, :year, step) do
+    %{year: year, month: month, day: day} = Cldr.Calendar.plus(naive_datetime, :years, step)
+    %NaiveDateTime{naive_datetime | year: year, month: month, day: day}
+  end
+
   def weeks_from_months(months, _config) when months == 0 do
     0
   end