+++ mix/backend.ex
@@ -3,10 +3,13 @@ require Cldr.Calendar.Backend.Compiler
 
 defmodule MyApp.Cldr do
   use Cldr,
-    providers: [Cldr.Calendar],
-    locales: ["en", "fr", "en-GB", "en-AU", "en-CA", "ar", "he"]
+    providers: [Cldr.Calendar, Cldr.Number, Cldr.Unit, Cldr.List],
+    locales: ["en", "fr", "en-GB", "en-AU", "en-CA", "ar", "he"],
+    default_locale: "en"
 end
 
 defmodule NoDocs.Cldr do
-  use Cldr, generate_docs: false, providers: [Cldr.Calendar]
+  use Cldr,
+    generate_docs: false,
+    providers: [Cldr.Calendar]
 end