+++ test/cldr_calendars_test.exs
@@ -1,6 +1,8 @@
 defmodule Cldr.Calendar.Test do
   use ExUnit.Case, asynch: true
   doctest Cldr.Calendar
+  doctest Cldr.Calendar.Duration
+  doctest Cldr.Calendar.Interval
   doctest Cldr.Calendar.Kday
 
   # :calendar module doesn't work with year 0 or negative years